+++ src/core/ui/components/sign-in/start.tsx
@@ -1,18 +1,18 @@
 import { withCardStateProvider } from '../../context/card';
 import { Flow } from '../flow/flow';
-import { SignInCard } from './card';
 import { useSignInContext } from './context';
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useCallback, useEffect, useState } from 'react';
 import { useRouter } from '../../router/context';
 import { useWeb3LoginState } from '../../context/web3-signin';
-import { LoadingModalContents } from './loading';
 import {
   LoginMethod,
   LoginMethodType,
   useLoginMethods,
 } from '../../context/login-methods';
-import { SignInButtons } from './sign-in-buttons';
 import { useInteraction } from '../../context/interaction';
+import { SignInScreen } from './screens/sign-in';
+import { LoadingScreen } from './screens/loading';
+import { useCoreSlashAuth } from '../../context/core-slashauth';
 
 type Props = {
   showAllWallets?: boolean;
@@ -21,15 +21,16 @@ type Props = {
 
 const _SignInStart = ({ showAllWallets, showBackButton }: Props) => {
   const { setProcessing } = useInteraction();
-  const { viewOnly, walletConnectOnly, connectAccounts } = useSignInContext();
+  const { viewOnly, walletConnectOnly } = useSignInContext();
   const { navigate, fullPath } = useRouter();
   const web3LoginState = useWeb3LoginState();
-  const { loginMethods, setSelectedLoginMethod } = useLoginMethods();
+  const { setSelectedLoginMethodById, selectedLoginMethod } = useLoginMethods();
+  const slashAuth = useCoreSlashAuth();
 
   const [isConnecting, setConnecting] = useState(false);
 
   useEffect(() => {
-    setSelectedLoginMethod(null);
+    setSelectedLoginMethodById(null);
     setConnecting(false);
     return () => {
       setConnecting(false);
@@ -37,25 +38,13 @@ const _SignInStart = ({ showAllWallets, showBackButton }: Props) => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
-  const headerText = useMemo(() => {
-    if (walletConnectOnly) {
-      return 'Continue with:';
-    }
-    if (connectAccounts) {
-      return 'Connect accounts with:';
-    }
-    return 'Sign in with:';
-  }, [connectAccounts, walletConnectOnly]);
-
   const handleLoginButtonClick = useCallback(
     async (loginMethod: LoginMethod) => {
       if (viewOnly) {
         return;
       }
       try {
-        setSelectedLoginMethod(
-          loginMethods.find((m) => m.id === loginMethod.id)
-        );
+        setSelectedLoginMethodById(loginMethod.id);
         setConnecting(true);
         if (loginMethod.type === LoginMethodType.Web3) {
           try {
@@ -91,10 +80,9 @@ const _SignInStart = ({ showAllWallets, showBackButton }: Props) => {
     },
     [
       fullPath,
-      loginMethods,
       navigate,
       setProcessing,
-      setSelectedLoginMethod,
+      setSelectedLoginMethodById,
       viewOnly,
       walletConnectOnly,
       web3LoginState.web3Manager,
@@ -103,44 +91,18 @@ const _SignInStart = ({ showAllWallets, showBackButton }: Props) => {
 
   return (
     <Flow.Part part="start">
-      <SignInCard showBackButton={showBackButton}>
-        {isConnecting ? (
-          <LoadingModalContents />
-        ) : (
-          <div
-            className="slashauth-modal-scrollable"
-            style={{
-              overflowY: 'hidden',
-              width: '100%',
-              marginTop: '2rem',
-            }}
-          >
-            <div
-              style={{
-                display: 'flex',
-                flexDirection: 'column',
-                overflowY: 'auto',
-              }}
-            >
-              <div>
-                <p
-                  style={{
-                    fontSize: '14px',
-                    fontWeight: 700,
-                    textAlign: 'start',
-                  }}
-                >
-                  {headerText}
-                </p>
-              </div>
-              <SignInButtons
-                onClick={handleLoginButtonClick}
-                showAllWallets={showAllWallets}
-              />
-            </div>
-          </div>
-        )}
-      </SignInCard>
+      {isConnecting ? (
+        <LoadingScreen
+          description={`Connecting to ${selectedLoginMethod.name}`}
+          detailedDescription={`Confirm your wallet connection in ${slashAuth.appName}`}
+          navigateBack={async () => {
+            await web3LoginState.web3Manager.disconnect();
+            setConnecting(false);
+          }}
+        />
+      ) : (
+        <SignInScreen startLoginWith={handleLoginButtonClick} />
+      )}
     </Flow.Part>
   );
 };