+++ src/core/ui/context/appearance.tsx
@@ -4,19 +4,29 @@ import { useDeepEqualMemo } from '../../../shared/hooks';
 import {
   ComputedSlashAuthModalStyle,
   SlashAuthStyle,
+  ThemeType,
 } from '../../../shared/types';
 import { createContextAndHook } from '../../../shared/utils';
-import { DARK_SLASHAUTH_ICON } from '../components/app-logo';
+import {
+  DARK_SLASHAUTH_ICON,
+  LIGHT_SLASHAUTH_ICON,
+} from '../components/app-logo';
 
 const DEFAULT_MODAL_STYLES: ComputedSlashAuthModalStyle = {
+  type: ThemeType.Light,
   backgroundColor: '#ffffff',
   buttonBackgroundColor: '#ffffff',
   hoverButtonBackgroundColor: '#f5f5f5',
-  borderRadius: '8px',
+  borderRadius: '22px',
   fontFamily: 'sans-serif',
   fontColor: '#000000',
   alignItems: 'center',
   iconURL: DARK_SLASHAUTH_ICON,
+  headerBackgroundColor: '#F3F4F6',
+  headerFontColor: '#363849',
+  lineColor: '#E5E7EB',
+  primaryButtonBackgroundColor: '#424559',
+  primaryButtonTextColor: 'white',
 };
 
 type AppearanceContextValue = {
@@ -30,12 +40,23 @@ type AppearanceProviderProps = React.PropsWithChildren<SlashAuthStyle>;
 
 const AppearanceProvider = (props: AppearanceProviderProps) => {
   const ctxValue = useDeepEqualMemo(() => {
+    const theme = {
+      ...DEFAULT_MODAL_STYLES,
+      ...props.signInModalStyle,
+    };
+
+    if (theme.type === ThemeType.Dark) {
+      theme.iconURL = LIGHT_SLASHAUTH_ICON;
+    }
+
+    const root = document.documentElement;
+    Object.keys(theme).forEach((key) => {
+      root.style.setProperty(`--slashauth-${key}`, theme[key]);
+    });
+
     return {
       value: {
-        modalStyle: {
-          ...DEFAULT_MODAL_STYLES,
-          ...props.signInModalStyle,
-        },
+        modalStyle: theme,
       },
     };
   }, [props.signInModalStyle]);